+++ src/main/java/baykov/daniel/springbootlibraryapp/services/BookService.java
@@ -7,18 +7,18 @@ public interface BookService {
 
     BookDTO createBook(BookDTO bookDTO);
 
-    BookDTO getBookById(long id);
+    BookDTO getBookById(Long id);
 
     BookResponse getAllBooks(int pageNo, int pageSize, String sortBy, String sortDir);
 
-    BookDTO updateBookByBookId(BookDTO bookDTO, long id);
+    BookDTO updateBookByBookId(BookDTO bookDTO, Long id);
 
-    void deleteBookByBookId(long id);
+    void deleteBookByBookId(Long id);
 
     void updateNumberOfBooksAfterBorrowing(Long id);
 
     void updateNumberOfBooksAfterReturning(Long id);
 
     BookResponse getBooksByTitleOrByGenreOrByDescriptionOrByPublicationYearOrByAuthorName(
-            String title, String genre, String description, int publicationYear, String authorFirstName, String authorLastName, int pageNo, int pageSize, String sortBy, String sortDir);
+            String title, String genre, String description, Long publicationYear, String authorFirstName, String authorLastName, int pageNo, int pageSize, String sortBy, String sortDir);
 }
